requirements.txt
asgiref==3.7.2
azure-core==1.29.3
azure-storage-blob==12.17.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
idna==3.4
isodate==0.6.1
mysqlclient==2.2.0
pycparser==2.21
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
